requirements.txt
streamlit~=1.48.0
pandas~=2.3.1
numpy~=2.2.6
python-dotenv~=1.1.1
pinecone~=7.3.0